requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
selenium==4.16.0
webdriver_manager==4.0.1 